fix: 불필요한코드 삭제 및 사용자 편의성 향상 (#40)
* feat: todo 삭제 전 confirm 추가

* fix: 불필요한 코드 삭제

* feat: todo 수정 전 confirm 추가

* feat: 로그아웃전 confirm 추가

+++ src/components/todo/Buttons/RemoveBtn/RemoveBtn.tsx
@@ -9,8 +9,9 @@ export default function RemoveBtn({ id }: { id: number }) {
 
   const handleDelete = useCallback(
     async (todoId: number) => {
-      await todoApi.deleteTodo(todoId);
+      if (!window.confirm("정말 삭제하시겠습니까?")) return;
 
+      await todoApi.deleteTodo(todoId);
       getTodos();
     },
     [getTodos]
